+++ src/core/interval.h
@@ -5,10 +5,10 @@
 namespace hasmet {
 class Interval {
  public:
-  double min, max;
+  float min, max;
 
   Interval() : min(INFINITY), max(-INFINITY) {};
-  Interval(double _min, double _max) : min(_min), max(_max) {}
+  Interval(float _min, float _max) : min(_min), max(_max) {}
   Interval(Interval _i0, Interval _i1)
       : min(std::min(_i0.min, _i1.min)), max(std::max(_i0.max, _i1.max)) {}
 
@@ -24,9 +24,9 @@ class Interval {
   inline bool overlap(const Interval& _other) const {
     return (min <= _other.max && _other.min <= max);
   }
-  inline bool consists(const double& point) const {
+  inline bool consists(const float& point) const {
     return (min <= point && max >= point);
   }
-  inline double getLength() const { return max - min; }
+  inline float getLength() const { return max - min; }
 };
 } // namespace hasmet
